AI_MCP/requirements.txt
boto3>=1.28.0
requests>=2.31.0
fastmcp>=0.1.0
pytest>=7.4.0 